requirements.txt
numpy
pandas
matplotlib
flask
flask_cors
torchvision 
torchaudio
scipy
scikit-learn
graspologic 
jupyterlab
pymupdf
tokenizers
transformers
sentence-transformers
datasets
accelerate
azure-identity
azure-storage-blob
peft
gensim
ollama
langchain
bitsandbytes
deepspeed
openai==1.58.1
tiktoken
dotenv
chunker
networkx
dotenv